requirements.txt
asgiref==3.5.0
astroid==2.9.3
attrs==21.4.0
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
filelock==3.6.0
identify==2.4.12
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
nodeenv==1.6.0
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
pylint==2.12.2
pymongo==3.12.3
pyparsing==3.0.7
pytest==7.1.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
virtualenv==20.13.4
wrapt==1.13.3
